[flutter tools] Move _informUserOfCrash into crash_reporting.dart
I intend to override `_informUserOfCrash` in google3.  Move it to
a new `CrashReporter` class and make it public in preparation for
this.

I think it's a bit confusing that there is a `CrashReporter` class
and a `CrashReportSender` class, but I haven't thought of a better
name, and there's been pushback on merging them.

This change is almost entirely code motion; the only change to the
moved code is to rename `_informUserOfCrash` to `informUser`.
